requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
pyyaml==6.0.1
python-telegram-bot==20.3
watchdog==3.0.0
imutils==0.5.4
scipy==1.11.2
ffmpeg-python==0.2.0

# Web dashboard
Flask==2.3.3
Flask-SocketIO==5.3.4
Flask-CORS==4.0.0
python-socketio==5.9.0
eventlet==0.33.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tzlocal==5.0.1
requests==2.31.0
psutil==5.9.5
colorlog==6.7.0

# Database (optional)
sqlalchemy==2.0.19
alembic==1.12.0

# Development
pytest==7.4.2
pytest-asyncio==0.21.1